sunxi: don't define MACPWR and SATAPWR to empty strings
There's checks in board/sunxi/board.c if either MACPWR or SATAPWR are
defined and they are defined by default to a empty string which means
on vast majority of AllWinner boards when they're not required the
code is still run and with SATAPWR we even get an unnecessary 500ms
delay booting. So let's not define a default for these options so the
code is only run for boards that need it.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>